llm_sorter/pyproject.toml
[tool.poetry]
name = "llm-sorter"
version = "0.1.0"
description = ""
authors = ["yessense <akorchemnyj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
wandb = "^0.15.12"
hydra-core = "^1.3.2"
numpy = "^1.26.1"
uvicorn = "^0.24.0.post1"
fastapi = "^0.104.1"
inflect = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
dvc = "^3.24.0"
sphinx = "^7.2.6"
numpydoc = "^1.6.0"
sphinx-rtd-theme = "^1.3.0"
pytest = "^7.4.2"

[tool.black]
line-length = 90
target-version = ["py310"]
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
    | .txt
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "llm_sorter"
testpaths = ["tests"]
filterwarnings = [
"ignore::DeprecationWarning"
]
